add bank service

+++ payment-service/src/main/java/com/coworking/space/paymentservice/domain/entities/PaymentEntity.java
@@ -30,5 +30,5 @@ public class PaymentEntity {
 
     private Instant pendingBoundary;
 
-    private int paymentAmount;
+    private double paymentAmount;
 }
